// src/app/enumerators/app.enum.ts
export enum Breackpoints {
	XS = 'xs',
	SM = 'sm',
	MD = 'md',
	LG = 'lg',
	XL = 'xl',
}

export enum Languages {
	ES = 'es',
	EN = 'en',
}

export enum Theme {
	LIGHT_THEME = 'light-theme',
	DARK_THEME = 'dark-theme',
}

export enum Categories {
	TOP = 'TOP',
	WEDDING = 'WEDDING',
	QUINCE = 'QUINCE',
	BABY = 'BABY',
	BOUDOIR = 'BOUDOIR',
	CHILDREN = 'CHILDREN',
	EVENTS = 'EVENTS',
	PRODUCTS = 'PRODUCTS',
	FASHION = 'FASHION',
}